sherpa_django/requirements.txt
Django==3.1.2
djangorestframework==3.11.1
requests==2.24.0
